+++ disropt/utils/__init__.py
@@ -1,2 +1,1 @@
-from .utilities import is_pos_def
-from .LexLPUtils import *
+from .utilities import *